# Changes current handling for replication lag in favor of setting lagging servers to SHUNNED state