--- a/include/MySQL_Thread.h
+++ b/include/MySQL_Thread.h
@@ -410,6 +410,7 @@ class MySQL_Threads_Handler
 		int monitor_groupreplication_healthcheck_timeout;
 		int monitor_groupreplication_healthcheck_max_timeout_count;
 		int monitor_groupreplication_max_transactions_behind_count;
+		int monitor_groupreplication_max_transactions_behind_for_read_only;
 		int monitor_galera_healthcheck_interval;
 		int monitor_galera_healthcheck_timeout;
 		int monitor_galera_healthcheck_max_timeout_count;