… include/proxysql_structs.h
… lib/MySQL_Thread.cpp
… src/SQLite3_Server.cpp