@bskllzh 2021-08-19
@JavierJF @renecannao , I think there are not many scenarios for mgr with multiple master.  For single master mode, there are more scenarios for using single master mode.  And when the slave server is in a state where the delay exceeds the threshold,  proxysql will immediately offline the slave server.  I think this is inappropriate, because it will interrupt the business and cause the program to report an error.  I submitted a fix PR, set it to OFFLINE_SOFT , and softly released the delay slave server.  Please review PR: #3473.