@renecannao 2021-08-25
Hi @bskllzh . Thank you for your feedback.
I think I get your point, and I absolutely agree with the problem you are pointing at.
Although, I think `OFFLINE_SOFT` is not the right approach.
Let me explain.

`OFFLINE_SOFT` is a configuration state, and a server in this state is *configured* to not be used for new connections, but not only...
`SHUNNED` is instead a *temporary status* from which the server should automatically recover.
In other words, a server shouldn't *automatically* go from `OFFLINE_SOFT` to `ONLINE` , but should *automatically* go from `SHUNNED` to `ONLINE`.

In fact, PR #3473 would conflict with what said previously: a server in `OFFLINE_SOFT` should never be returned to `ONLINE` automatically (and this is now implemented in PR #3533).

> And when the slave server is in a state where the delay exceeds the threshold, proxysql will immediately offline the slave server. I think this is inappropriate, because it will interrupt the business and cause the program to report an error.

This is by design.
All the details are here: #774
We could set status to `MYSQL_SERVER_STATUS_SHUNNED` instead of `MYSQL_SERVER_STATUS_SHUNNED_REPLICATION_LAG` .
This will solve the issue of connections being closed immediately, but hostgroup manager automatically tries to bring a server back online from shunned, no matter if replication lag is still present or not: this is why `MYSQL_SERVER_STATUS_SHUNNED_REPLICATION_LAG` exists and is different than state `MYSQL_SERVER_STATUS_SHUNNED` .

Thinking about a possible solution, we could implement a mechanism in which a node is first configured as `MYSQL_SERVER_STATUS_SHUNNED` and then `MYSQL_SERVER_STATUS_SHUNNED_REPLICATION_LAG` if replication lag doesn't recover quickly.
The state `MYSQL_SERVER_STATUS_SHUNNED_REPLICATION_LAG` should be set within a short period of time after `MYSQL_SERVER_STATUS_SHUNNED` because otherwise the node could go `ONLINE` while it shouldn't.